The CSS File
Here we did many things, we set the extension width and position. We also ajusted the alignment of its elments. Then we focused on the section containing the picker and the picker itself and we gave it a hoover effect. We also worked on the second section (positioning, padding, alignment, space-between, etc) but we also hid it for now, we are going to work on that later on with the javascript.