The Javascript file
Finally, this one was the hardest to get. You need first to get all your section and access to the storage. Then you create a function that displays the picked colors and that will be called anytime a color gets picked. Then you need to actually pick the colors via the color picker trough another function the tutorial helped us greatly with and finally you have to create the Clear function which was the easiest to do